Add favourite and likes

+++ app/controllers/products_controller.rb
@@ -1,5 +1,5 @@
 class ProductsController < ApplicationController
-  # before_action :authenticate_user!, except: [:index, :show]
+  before_action :authenticate_user!, except: [:index, :show]
 
   def new
     @product = Product.new
@@ -15,6 +15,7 @@ class ProductsController < ApplicationController
     @product.user = current_user
     if @product.save
       flash[:notice] = "New product created"
+      ProductsMailer.notify_product_owner(@product).deliver_now
       redirect_to product_path(@product)
     else
       render :new
@@ -23,30 +24,30 @@ class ProductsController < ApplicationController
 
   def show
     @product = Product.find params[:id]
-    # @review = Review.new
+    @review = Review.new
   end
 
   def edit
     @product = Product.find params[:id]
-  #   redirect_to root_path, alert:'access denied' unless can? :edit, @product
+    redirect_to root_path, alert:'access denied' unless can? :edit, @product
   end
 
   def update
 
     @product = Product.find params[:id]
     product_params = params.require(:product).permit([:title, :description, :price, :category_id])
-    if @product.update(product_params)
-      redirect_to product_path(@product)
-    else
-      render :edit
-    end
-    # if !(can? :edit, @product)
-    #   redirect_to root_path, alert:'access denied'
-    # elsif @product.update(product_params)
+    # if @product.update(product_params)
     #   redirect_to product_path(@product)
     # else
     #   render :edit
     # end
+    if !(can? :edit, @product)
+      redirect_to root_path, alert:'access denied'
+    elsif @product.update(product_params)
+      redirect_to product_path(@product)
+    else
+      render :edit
+    end
   #   # if @product.user != current_user
   #   #   flash[:alert] = "You cannot change a product that you did not create"
   #   #   redirect_to product_path(@product)
@@ -59,13 +60,13 @@ class ProductsController < ApplicationController
   end
 
   def destroy
-    # if !(can? :destroy, @product)
-    #   redirect_to root_path, alert:'access denied'
-    # else
+    if !(can? :destroy, @product)
+      redirect_to root_path, alert:'access denied'
+    else
       product = Product.find params[:id]
       product.destroy
-      # redirect_to products_path
-    # end
+      redirect_to products_path
+    end
   end
 
 end
